pyproject.toml
[tool.poetry]
name = "sorrel"
version = "1.3.0"
description = "A flexible framework for developing reinforcement learning algorithms."
license = "MIT"
readme = "README.md"
maintainers = [
    "Yibing Ju <bing.ju@mail.utoronto.ca>", 
    "Rebekah Gelpi <rebekah.gelpi@mail.utoronto.ca>"
]
authors = [
    "Ethan Jackson <jackson.ethan.c@gmail.com>",
    "Shon Verch <verchshon@gmail.com>",
    "Yikai Tang <yikai.tang@mail.utoronto.ca>",
    "Claas Voelcker <claas@voelcker.net>",
    "Wil Cunningham"
]
packages = [{ include = "sorrel" }]
homepage = "https://github.com/social-ai-uoft/sorrel"
documentation = "https://sorrel.readthedocs.io"

[tool.poetry.urls]
Paper = "https://doi.org/10.48550/arXiv.2506.00228"

[tool.poetry.dependencies]
python = "^3.12.0"
matplotlib = "^3.6.1"
numpy = "^1.26.4"
pyyaml = "^6.0.2"
pytest = "^8.3.5"

torch = "^2.7.0"
jax = "^0.4.35"
flax = "^0.10.1"

# used for configs
hydra-core = "^1.3.0"

# used in buffer.py
dill = "^0.3.9"
# used in human_player.py
ipython = "^8.29.0"
# used in utils.py
pillow = "^11.0.0"
# used in logging.py
tensorboard = "^2.18.0"
openai = "^1.102.0"
pygame = "^2.6.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

# for docs
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"
myst-nb = "^1.1.2"
sphinx-autodoc2 = "^0.5.0"

[tool.pytest.ini_options]
pythonpath = [
  "sorrel"
]
addopts = [
  "--doctest-modules", 
  "--import-mode=importlib"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
